+++ NEWS.md
@@ -8,6 +8,7 @@
 
   * amTimesSeries : add connect options + date format
   * getTransformTS : add col_by possibility + optimize performance
+  * getTransformTS : fix data.table on iTime with some tz
   
 ### rAmCharts 2.1.12 (based on [amcharts][amcharts_url] version **3.21.13**)